+++ test/kubernetes/e2e/features/headless_svc/inputs/classic_api.gen.yaml
