+++ test/kubernetes/e2e/features/headless_svc/suite.go
@@ -3,13 +3,15 @@ package headless_svc
 import (
 	"context"
 
+	"github.com/stretchr/testify/suite"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
 	"github.com/solo-io/gloo/pkg/utils/kubeutils"
 	"github.com/solo-io/gloo/pkg/utils/requestutils/curl"
 	"github.com/solo-io/gloo/projects/gateway/pkg/defaults"
 	"github.com/solo-io/gloo/test/kubernetes/e2e"
-	"github.com/stretchr/testify/suite"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
+	"github.com/solo-io/gloo/test/kubernetes/e2e/utils"
 )
 
 type headlessSvcSuite struct {
@@ -34,18 +36,16 @@ func NewHeadlessSvcTestingSuite(ctx context.Context, testInst *e2e.TestInstallat
 
 // SetupSuite generates manifest files for the test suite
 func (s *headlessSvcSuite) SetupSuite() {
-
 	if s.useK8sGatewayApi {
 		// use the k8s gateway api resources
 		resources := []client.Object{gw, headlessSvcHTTPRoute}
-		err := writeResourcesToFile(resources, k8sApiRoutingManifest)
+		err := utils.WriteResourcesToFile(resources, k8sApiRoutingManifest)
 		s.Require().NoError(err, "can write resources to file")
 	} else {
 		resources := getClassicEdgeResources(s.testInstallation.Metadata.InstallNamespace)
-		err := writeResourcesToFile(resources, classicApiRoutingManifest)
+		err := utils.WriteResourcesToFile(resources, classicApiRoutingManifest)
 		s.Require().NoError(err, "can write resources to file")
 	}
-
 }
 
 func (s *headlessSvcSuite) TestConfigureRoutingHeadlessSvc() {
@@ -99,5 +99,4 @@ func (s *headlessSvcSuite) TestConfigureRoutingHeadlessSvc() {
 			},
 			expectedHealthyResponse)
 	}
-
 }
